# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API for Solana Memecoin Intelligence
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type OverviewMetricTrend = typeof OverviewMetricTrend[keyof typeof OverviewMetricTrend];


export const OverviewMetricTrend = {
  up: 'up',
  down: 'down',
  flat: 'flat',
} as const;

export interface OverviewMetric {
  label: string;
  value: string;
  change: number;
  trend: OverviewMetricTrend;
}

export type TokenRiskLabel = typeof TokenRiskLabel[keyof typeof TokenRiskLabel];


export const TokenRiskLabel = {
  low: 'low',
  guarded: 'guarded',
  elevated: 'elevated',
  critical: 'critical',
} as const;

export interface TokenRisk {
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  label: TokenRiskLabel;
  liquidityLocked: boolean;
  mintAuthorityActive: boolean;
  freezeAuthorityActive: boolean;
  top10HolderPct: number;
  notes: string[];
}

export interface Token {
  address: string;
  symbol: string;
  name: string;
  /** @nullable */
  icon: string | null;
  price: number;
  priceChange24h: number;
  volume24h: number;
  liquidity: number;
  marketCap: number;
  txns24h: number;
  ageMinutes: number;
  dex: string;
  risk: TokenRisk;
  /**
     * @minimum 0
     * @maximum 100
     */
  momentum: number;
  isWatched: boolean;
  sparkline: number[];
}

export type ActivityEventType = typeof ActivityEventType[keyof typeof ActivityEventType];


export const ActivityEventType = {
  launch: 'launch',
  liquidity: 'liquidity',
  whale: 'whale',
  risk: 'risk',
  social: 'social',
} as const;

export type ActivityEventSeverity = typeof ActivityEventSeverity[keyof typeof ActivityEventSeverity];


export const ActivityEventSeverity = {
  info: 'info',
  positive: 'positive',
  warning: 'warning',
  critical: 'critical',
} as const;

export interface ActivityEvent {
  id: string;
  type: ActivityEventType;
  title: string;
  detail: string;
  tokenSymbol: string;
  severity: ActivityEventSeverity;
  createdAt: string;
}

export interface MarketOverview {
  metrics: OverviewMetric[];
  tokens: Token[];
  events: ActivityEvent[];
  generatedAt: string;
}

export interface WatchlistItem {
  address: string;
  symbol: string;
  name: string;
  addedAt: string;
}

export interface WatchlistInput {
  address: string;
  symbol: string;
  name: string;
}

export type AlertRuleCondition = typeof AlertRuleCondition[keyof typeof AlertRuleCondition];


export const AlertRuleCondition = {
  price_change: 'price_change',
  liquidity_drop: 'liquidity_drop',
  risk_score: 'risk_score',
} as const;

export type AlertRuleDirection = typeof AlertRuleDirection[keyof typeof AlertRuleDirection];


export const AlertRuleDirection = {
  above: 'above',
  below: 'below',
} as const;

export interface AlertRule {
  id: number;
  address: string;
  symbol: string;
  condition: AlertRuleCondition;
  threshold: number;
  direction: AlertRuleDirection;
  enabled: boolean;
  createdAt: string;
}

export type AlertRuleInputCondition = typeof AlertRuleInputCondition[keyof typeof AlertRuleInputCondition];


export const AlertRuleInputCondition = {
  price_change: 'price_change',
  liquidity_drop: 'liquidity_drop',
  risk_score: 'risk_score',
} as const;

export type AlertRuleInputDirection = typeof AlertRuleInputDirection[keyof typeof AlertRuleInputDirection];


export const AlertRuleInputDirection = {
  above: 'above',
  below: 'below',
} as const;

export interface AlertRuleInput {
  address: string;
  symbol: string;
  condition: AlertRuleInputCondition;
  threshold: number;
  direction: AlertRuleInputDirection;
}

export type TokenAnalysisVerdict = typeof TokenAnalysisVerdict[keyof typeof TokenAnalysisVerdict];


export const TokenAnalysisVerdict = {
  watch: 'watch',
  investigate: 'investigate',
  avoid: 'avoid',
} as const;

export type TokenAnalysisSignalsItemStatus = typeof TokenAnalysisSignalsItemStatus[keyof typeof TokenAnalysisSignalsItemStatus];


export const TokenAnalysisSignalsItemStatus = {
  positive: 'positive',
  neutral: 'neutral',
  negative: 'negative',
} as const;

export type TokenAnalysisSignalsItem = {
  label: string;
  status: TokenAnalysisSignalsItemStatus;
  detail: string;
};

export interface TokenAnalysis {
  address: string;
  symbol: string;
  verdict: TokenAnalysisVerdict;
  /**
     * @minimum 0
     * @maximum 100
     */
  confidence: number;
  summary: string;
  signals: TokenAnalysisSignalsItem[];
  generatedAt: string;
}

export type ListTokensParams = {
sort?: ListTokensSort;
/**
 * @nullable
 */
search?: string | null;
/**
 * @minimum 1
 * @maximum 20
 */
limit?: number;
};

export type ListTokensSort = typeof ListTokensSort[keyof typeof ListTokensSort];


export const ListTokensSort = {
  momentum: 'momentum',
  volume: 'volume',
  newest: 'newest',
  risk: 'risk',
} as const;